Finished and tested command creation and handling
The `/ping` command code was completed (along with a modification of its data) and was removed `/ping-test` command.
Added condition to update commands data if needed manageable in the `.env` file.
Updated also `BlueCommand` class to solve minor bugs when dealing with undefined values.
Lastly was removed a debug logging printed when triggering command events.