Replace translation buttons with dropdown menu
Changes the translation method: instead of one button for each language it puts one single drop sown select menu with all the supported languages.
It also de-centralizes the `getVariables()` function in order to be used both for embed and message translations.

NOTE: The interaction generated by the string select menu it's still not being processed properly.